Refactor to use useSelectedBlockToolProps
